use float as time

+++ include/reactive/Timer/CPUTimer.hpp
@@ -8,13 +8,13 @@ public:
 
     void restart() { startTime = std::chrono::steady_clock::now(); }
 
-    auto elapsedInNano() const -> double {
+    auto elapsedInNano() const -> float {
         auto endTime = std::chrono::steady_clock::now();
-        return static_cast<double>(
+        return static_cast<float>(
             std::chrono::duration_cast<std::chrono::nanoseconds>(endTime - startTime).count());
     }
 
-    auto elapsedInMilli() const -> double { return elapsedInNano() / 1000000.0; }
+    auto elapsedInMilli() const -> float { return elapsedInNano() / 1000000.0f; }
 
 private:
     std::chrono::time_point<std::chrono::steady_clock> startTime;
